// src/app/pages/Models/InformacionComplementariaDto.ts
// InformacionComplementariaDto.ts
export interface InformacionComplementariaDto {
    id: number;
    idFormulario: number;
    activosVirtuales: boolean;
    grandesCantidadesEfectivo: boolean;
    investigadoViolacionLeyesAnticorrupcion: boolean;
    declaracionNoToleranciaCorrupcion: boolean;
    extensionColaboradoresPolitica: boolean;
    politicaAportesDonaciones: boolean;
    contratadoTercerosOrganizacion: boolean;
    obligadaSistemaPrevencionLAFT: boolean;
    tieneSistemaPrevencionLAFT: boolean;
    adopcionPoliticasLAFT: boolean;
    nombramientoOficialCumplimiento: boolean;
    medidasDebidaDiligencia: boolean;
    identificacionEvaluacionRiesgos: boolean;
    politicasCapacitacionLAFT: boolean;
    obligadoAutocontrolLAFT: boolean;
    obligadoProgramaPTEE: boolean;
    adopcionPoliticasOrganoDireccion: boolean;
    establecimientoMedidasDebidaDiligencia: boolean;
    identificacionReportesOperSospechosas: boolean;
    riesgosCorrupcionSobornoTransnacional: boolean;
    riesgosLAFT: boolean;
    politicasCapacitacion: boolean;
  }
  